refactor(vt-daemon): group spawn-time *Injection helpers into spawn/injection/
Adding ensureHomePrompts.ts pushed packages/.../agent-runtime/spawn past the
15-child directory-fanout gate (16 children). Rather than raise the limit
(reward-hacking the gate), group the three cohesive spawn-time injection
helpers — agentHookInjection, cliManualInjection, vtPathInjection (shared
'*Injection' purpose: inject hook flags / CLI manual / vt PATH into the spawned
agent) — into spawn/injection/. Brings the dir to 14 children with headroom.

Pure path-move + import-rewrite; no behavior change. All spawn tests green.
