+++ packages/systems/vt-daemon/src/agent-runtime/spawn/buildTerminalEnvVars.ts
@@ -8,8 +8,8 @@ import type {VTSettings} from '@vt/graph-model/settings'
 import {getRuntimeEnv} from '../runtime/runtime-config'
 import {getProjectDotVoicetreePath, resolveVoicetreeHomePath} from '@vt/paths'
 import {getRuntimeProjectRoot, getRuntimeProjectPaths} from '../runtime/graph-bridge'
-import {appendCliManualToAgentPrompt} from './cliManualInjection'
-import {prependVtBinToPath, readVtBinDirOrNull} from './vtPathInjection'
+import {appendCliManualToAgentPrompt} from './injection/cliManualInjection'
+import {prependVtBinToPath, readVtBinDirOrNull} from './injection/vtPathInjection'
 import {readDaemonPortFromProject} from './daemonUrlFile'
 import {promises as fs} from 'fs'
 import type {Dirent} from 'fs'
