+++ packages/systems/vt-daemon/src/agent-runtime/spawn/injection/agentHookInjection.ts
