+++ packages/systems/vt-daemon/src/agent-runtime/spawn/injection/cliManualInjection.ts
