+++ packages/systems/vt-daemon/src/agent-runtime/spawn/terminalData.ts
@@ -6,7 +6,7 @@ import {getNextAgentName, getUniqueAgentName} from '@vt/graph-model/settings'
 import {createTerminalData, type TerminalData, type TerminalId} from '@vt/vt-daemon/agent-runtime/terminals/terminal-registry/types.ts'
 import {getExistingAgentNames} from '@vt/vt-daemon/agent-runtime/terminals/terminal-registry/index.ts'
 import {buildTerminalEnvVars} from './buildTerminalEnvVars'
-import {injectClaudeSettingsFlag, injectCodexHookFlags} from './agentHookInjection'
+import {injectClaudeSettingsFlag, injectCodexHookFlags} from './injection/agentHookInjection'
 import {ensureClaudeHookSettingsFile} from './claudeHookSettingsBootstrap'
 import {readDaemonPortFromProject} from './daemonUrlFile'
 import {getRuntimeEnv} from '../runtime/runtime-config'
