+++ packages/systems/vt-daemon/src/agent-runtime/spawn/tests/claudeHookSettingsBootstrap.test.ts
@@ -8,7 +8,7 @@
 
 import {describe, it, expect, beforeEach} from 'vitest'
 import {ensureClaudeHookSettingsFile, type ClaudeHookBootstrapDeps} from '../claudeHookSettingsBootstrap'
-import {buildClaudeHookSettingsJson} from '../agentHookInjection'
+import {buildClaudeHookSettingsJson} from '../injection/agentHookInjection'
 
 function makeInMemoryFs(initial: Record<string, string> = {}): {
     deps: ClaudeHookBootstrapDeps
